Create Top_EV_Cars
This repository contains an analysis of electric vehicle (EV) data from the my-project-1-366602.ev_cars.ev_dataset table in Google BigQuery. The goal is to identify the top 5 EV cars based on specific criteria, such as range or popularity. The repository includes an SQL query used to extract and analyze the data, along with a CSV file containing the query results. These resources can be used by researchers, enthusiasts, or decision-makers to gain insights into the EV market and compare different EV models.